refactoring and file structure update
