+++ src/context/CreateAccountFormikContext.ts
@@ -1,25 +1,6 @@
 import React from 'react';
 import {FormikProps} from 'formik';
-
-export type FormikInitialValues = {
-  firstName: string;
-  lastName: string;
-  email: string;
-  telephone: string;
-  address: Address;
-  avatar: string;
-  password: string;
-  confirmPassword: string;
-};
-
-type Address = {
-  street: string;
-  number: string;
-  city: string;
-  state: string;
-  postalCode: string;
-  country: string;
-};
+import {FormikInitialValues} from '../types/CreateAccountTypes';
 
 export const FormikContext =
   React.createContext<FormikProps<FormikInitialValues> | null>(null);
